requirements.txt
requests
beautifulsoup4
pillow>=10.3
pyyaml
pandas
ultralytics>=8.2
rembg>=2.0.57
onnxruntime